Move the traffic client code into its own file.
